+++ src/main/java/searchengine/repositories/PageRepository.java
@@ -2,7 +2,6 @@ package searchengine.repositories;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import searchengine.model.Page;
 import searchengine.model.Site;
@@ -11,7 +10,7 @@ import java.util.List;
 
 @Repository
 public interface PageRepository extends JpaRepository<Page, Integer> {
-    long countBySite(Site site);
+    int countBySite(Site site);
 
     Page findBySiteAndPath(Site site, String path);
